+++ linuxkm/linuxkm_wc_port.h
@@ -333,6 +333,13 @@
 
     #if defined(WOLFSSL_LINUXKM_USE_SAVE_VECTOR_REGISTERS) && \
         defined(CONFIG_X86)
+
+        extern __must_check int allocate_wolfcrypt_linuxkm_fpu_states(void);
+        extern void free_wolfcrypt_linuxkm_fpu_states(void);
+        extern __must_check int can_save_vector_registers_x86(void);
+        extern __must_check int save_vector_registers_x86(void);
+        extern void restore_vector_registers_x86(void);
+
         #if LINUX_VERSION_CODE < KERNEL_VERSION(4, 0, 0)
             #include <asm/i387.h>
         #else
@@ -368,8 +375,30 @@
         #ifndef RESTORE_VECTOR_REGISTERS
             #define RESTORE_VECTOR_REGISTERS() restore_vector_registers_x86()
         #endif
+
     #elif defined(WOLFSSL_LINUXKM_USE_SAVE_VECTOR_REGISTERS) && (defined(CONFIG_ARM) || defined(CONFIG_ARM64))
+
+        #error kernel module ARM SIMD is not yet tested or usable.
+
         #include <asm/fpsimd.h>
+
+        static WARN_UNUSED_RESULT inline int save_vector_registers_arm(void)
+        {
+            preempt_disable();
+            if (! may_use_simd()) {
+                preempt_enable();
+                return BAD_STATE_E;
+            } else {
+                fpsimd_preserve_current_state();
+                return 0;
+            }
+        }
+        static inline void restore_vector_registers_arm(void)
+        {
+            fpsimd_restore_current_state();
+            preempt_enable();
+        }
+
         #ifndef SAVE_VECTOR_REGISTERS
             #define SAVE_VECTOR_REGISTERS(fail_clause) { int _svr_ret = save_vector_registers_arm(); if (_svr_ret != 0) { fail_clause } }
         #endif
@@ -382,9 +411,10 @@
         #ifndef RESTORE_VECTOR_REGISTERS
             #define RESTORE_VECTOR_REGISTERS() restore_vector_registers_arm()
         #endif
+
     #elif defined(WOLFSSL_LINUXKM_USE_SAVE_VECTOR_REGISTERS)
         #error WOLFSSL_LINUXKM_USE_SAVE_VECTOR_REGISTERS is set for an unsupported architecture.
-    #endif
+    #endif /* WOLFSSL_LINUXKM_USE_SAVE_VECTOR_REGISTERS */
 
     _Pragma("GCC diagnostic pop");
 
@@ -529,39 +559,15 @@
         #endif
 
         struct task_struct *(*get_current)(void);
-        int (*preempt_count)(void);
 
         #ifdef WOLFSSL_LINUXKM_USE_SAVE_VECTOR_REGISTERS
 
-            #if LINUX_VERSION_CODE < KERNEL_VERSION(6, 2, 0)
-                typeof(cpu_number) *cpu_number;
-            #else
-                typeof(pcpu_hot) *pcpu_hot;
-            #endif
-            typeof(nr_cpu_ids) *nr_cpu_ids;
-
-            #if defined(CONFIG_SMP) && (LINUX_VERSION_CODE >= KERNEL_VERSION(5, 7, 0)) && !defined(WOLFSSL_COMMERCIAL_LICENSE)
-                /* note the current and needed version of these were added in af449901b8 (2020-Sep-17) */
-                typeof(migrate_disable) *migrate_disable;
-                typeof(migrate_enable) *migrate_enable;
-            #endif
-
             #ifdef CONFIG_X86
-                typeof(irq_fpu_usable) *irq_fpu_usable;
-                #ifdef WOLFSSL_COMMERCIAL_LICENSE
-                    typeof(fpregs_lock) *fpregs_lock;
-                    typeof(fpregs_lock) *fpregs_unlock;
-                #else /* !WOLFSSL_COMMERCIAL_LICENSE */
-                    /* kernel_fpu_begin() replaced by kernel_fpu_begin_mask() in commit e4512289,
-                     * released in kernel 5.11, backported to 5.4.93
-                     */
-                    #ifdef kernel_fpu_begin
-                        typeof(kernel_fpu_begin_mask) *kernel_fpu_begin_mask;
-                    #else
-                        typeof(kernel_fpu_begin) *kernel_fpu_begin;
-                    #endif
-                    typeof(kernel_fpu_end) *kernel_fpu_end;
-                #endif /* !defined(WOLFSSL_COMMERCIAL_LICENSE) */
+                typeof(allocate_wolfcrypt_linuxkm_fpu_states) *allocate_wolfcrypt_linuxkm_fpu_states;
+                typeof(can_save_vector_registers_x86) *can_save_vector_registers_x86;
+                typeof(free_wolfcrypt_linuxkm_fpu_states) *free_wolfcrypt_linuxkm_fpu_states;
+                typeof(restore_vector_registers_x86) *restore_vector_registers_x86;
+                typeof(save_vector_registers_x86) *save_vector_registers_x86;
             #else /* !CONFIG_X86 */
                 #error WOLFSSL_LINUXKM_USE_SAVE_VECTOR_REGISTERS is set for an unsupported architecture.
             #endif /* arch */
@@ -697,38 +703,15 @@
 
     #undef get_current
     #define get_current (wolfssl_linuxkm_get_pie_redirect_table()->get_current)
-    #undef preempt_count
-    #define preempt_count (wolfssl_linuxkm_get_pie_redirect_table()->preempt_count)
 
-    #ifdef WOLFSSL_LINUXKM_USE_SAVE_VECTOR_REGISTERS
-        #if LINUX_VERSION_CODE < KERNEL_VERSION(6, 2, 0)
-            #define cpu_number (*(wolfssl_linuxkm_get_pie_redirect_table()->cpu_number))
-        #else
-            #define pcpu_hot (*(wolfssl_linuxkm_get_pie_redirect_table()->pcpu_hot))
-        #endif
-        #define nr_cpu_ids (*(wolfssl_linuxkm_get_pie_redirect_table()->nr_cpu_ids))
-
-        #if defined(CONFIG_SMP) && (LINUX_VERSION_CODE >= KERNEL_VERSION(5, 7, 0)) && !defined(WOLFSSL_COMMERCIAL_LICENSE)
-            #define migrate_disable (*(wolfssl_linuxkm_get_pie_redirect_table()->migrate_disable))
-            #define migrate_enable (*(wolfssl_linuxkm_get_pie_redirect_table()->migrate_enable))
-        #endif
-
-        #ifdef CONFIG_X86
-            #define irq_fpu_usable (wolfssl_linuxkm_get_pie_redirect_table()->irq_fpu_usable)
-            #ifdef WOLFSSL_COMMERCIAL_LICENSE
-                #define fpregs_lock() (wolfssl_linuxkm_get_pie_redirect_table()->fpregs_lock())
-                #define fpregs_unlock() (wolfssl_linuxkm_get_pie_redirect_table()->fpregs_unlock())
-            #else /* !defined(WOLFSSL_COMMERCIAL_LICENSE) */
-                #ifdef kernel_fpu_begin
-                    #define kernel_fpu_begin_mask (wolfssl_linuxkm_get_pie_redirect_table()->kernel_fpu_begin_mask)
-                #else
-                    #define kernel_fpu_begin (wolfssl_linuxkm_get_pie_redirect_table()->kernel_fpu_begin)
-                #endif
-                #define kernel_fpu_end (wolfssl_linuxkm_get_pie_redirect_table()->kernel_fpu_end)
-            #endif /* !defined(WOLFSSL_COMMERCIAL_LICENSE) */
-        #else /* !CONFIG_X86 */
-            #error WOLFSSL_LINUXKM_USE_SAVE_VECTOR_REGISTERS is set for an unsupported architecture.
-        #endif /* archs */
+    #if defined(WOLFSSL_LINUXKM_USE_SAVE_VECTOR_REGISTERS) && defined(CONFIG_X86)
+        #define allocate_wolfcrypt_linuxkm_fpu_states (wolfssl_linuxkm_get_pie_redirect_table()->allocate_wolfcrypt_linuxkm_fpu_states)
+        #define can_save_vector_registers_x86 (wolfssl_linuxkm_get_pie_redirect_table()->can_save_vector_registers_x86)
+        #define free_wolfcrypt_linuxkm_fpu_states (wolfssl_linuxkm_get_pie_redirect_table()->free_wolfcrypt_linuxkm_fpu_states)
+        #define restore_vector_registers_x86 (wolfssl_linuxkm_get_pie_redirect_table()->restore_vector_registers_x86)
+        #define save_vector_registers_x86 (wolfssl_linuxkm_get_pie_redirect_table()->save_vector_registers_x86)
+    #elif defined(WOLFSSL_LINUXKM_USE_SAVE_VECTOR_REGISTERS)
+        #error WOLFSSL_LINUXKM_USE_SAVE_VECTOR_REGISTERS is set for an unsupported architecture.
     #endif /* WOLFSSL_LINUXKM_USE_SAVE_VECTOR_REGISTERS */
 
     #define __mutex_init (wolfssl_linuxkm_get_pie_redirect_table()->__mutex_init)
@@ -761,41 +744,6 @@
     #endif /* __PIE__ */
 
     #endif /* USE_WOLFSSL_LINUXKM_PIE_REDIRECT_TABLE */
-
-#ifdef WOLFSSL_LINUXKM_USE_SAVE_VECTOR_REGISTERS
-
-#ifdef CONFIG_X86
-
-    extern __must_check int allocate_wolfcrypt_linuxkm_fpu_states(void);
-    extern void free_wolfcrypt_linuxkm_fpu_states(void);
-    extern __must_check int can_save_vector_registers_x86(void);
-    extern __must_check int save_vector_registers_x86(void);
-    extern void restore_vector_registers_x86(void);
-
-#elif defined(CONFIG_ARM) || defined(CONFIG_ARM64)
-
-    #error kernel module ARM SIMD is not yet tested or usable.
-
-    static WARN_UNUSED_RESULT inline int save_vector_registers_arm(void)
-    {
-        preempt_disable();
-        if (! may_use_simd()) {
-            preempt_enable();
-            return BAD_STATE_E;
-        } else {
-            fpsimd_preserve_current_state();
-            return 0;
-        }
-    }
-    static inline void restore_vector_registers_arm(void)
-    {
-        fpsimd_restore_current_state();
-        preempt_enable();
-    }
-
-#endif
-
-#endif /* WOLFSSL_LINUXKM_USE_SAVE_VECTOR_REGISTERS */
 
     /* remove this multifariously conflicting macro, picked up from
      * Linux arch/<arch>/include/asm/current.h.
